Added current limit and ramp rate

+++ RobotCode2019/src/main/java/frc/robot/Drive.java
@@ -42,7 +42,7 @@ public class Drive {
         
         SmartDashboard.putNumber("rampRate", 0);
 
-        (int)SmartDashboard.putNumber("currentLimit", 0);
+        SmartDashboard.putNumber("currentLimit", 0);
     }
 
     /**
